[MOD] SignUp.jsx -> component 쪼개기 실패. 시간 관계 상 쪼개지 않고 기능 구현에 집중
