--- src/components/SignUpComponents/Sidebar.jsx
@@ -1,6 +1,0 @@
-import { AuthSidebar } from "./styled-components/SignUpStyle";
-
-function Sidebar() {
-  return <AuthSidebar />;
-}
-export default Sidebar;
