--- src/context/SignUpNavigateContext.jsx
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-
-export const SignUpNavigateContext = createContext(null);
-export const SignUpModeContext = createContext(null);
-export const SignUpDispatchContext = createContext(null);
-export const SignUpReduxActionContext = createContext(null);
